requirements.txt
# Notes RAG — Python dependencies
# Install: pip install -r requirements.txt

# Core framework
fastapi==0.135.1
uvicorn==0.41.0

# LangChain ecosystem
langchain==1.2.10
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-core==1.2.17
langchain-openai==1.1.10
langchain-text-splitters==1.1.1

# Embeddings + vector store
chromadb==1.5.4
fastembed>=0.4.0

# BM25 keyword search (now via SQLite FTS5 — no extra deps)

# Research pipeline
beautifulsoup4==4.14.3
requests==2.32.5

# File watching
watchdog==6.0.0

# Config
PyYAML==6.0.3

# Vault integration (optional)
hvac==2.4.0

# Tokenizer
tiktoken==0.12.0
